feat(pattern_assembler): Further broke out files
